moved api endpoints to service folder
