+++ agenta-web/src/components/Evaluations/AutomaticEvaluationResult.tsx
@@ -1,4 +1,8 @@
-import {deleteEvaluations, fetchEvaluationResults, loadEvaluations} from "@/lib/services/api"
+import {
+    deleteEvaluations,
+    fetchEvaluationResults,
+    fetchAllLoadEvaluations,
+} from "@/services/human-evaluations"
 import {Button, Spin, Statistic, Table, Typography} from "antd"
 import {useRouter} from "next/router"
 import {useEffect, useState} from "react"
@@ -101,7 +105,7 @@ export default function AutomaticEvaluationResult({
         const fetchEvaluations = async () => {
             try {
                 setFetchingEvaluations(true)
-                const evals: Evaluation[] = (await loadEvaluations(app_id)).map(
+                const evals: Evaluation[] = (await fetchAllLoadEvaluations(app_id)).map(
                     fromEvaluationResponseToEvaluation,
                 )
                 const results = await Promise.all(evals.map((e) => fetchEvaluationResults(e.id)))
