+++ agenta-web/src/components/Evaluations/HumanEvaluationResult.tsx
@@ -1,4 +1,5 @@
-import {deleteEvaluations, fetchData} from "@/lib/services/api"
+import {fetchData} from "@/services/api"
+import {deleteEvaluations} from "@/services/human-evaluations"
 import {Button, Spin, Statistic, Table, Typography} from "antd"
 import {useRouter} from "next/router"
 import {useEffect, useState} from "react"
