+++ agenta-web/src/components/HumanEvaluationModal/HumanEvaluationModal.tsx
@@ -1,6 +1,8 @@
 import React, {useEffect, useState} from "react"
 import {GenericObject, JSSTheme, Parameter, Variant} from "@/lib/Types"
-import {createNewEvaluation, fetchVariants, useLoadTestsetsList} from "@/lib/services/api"
+import {fetchVariants} from "@/services/api"
+import {createNewEvaluation} from "@/services/human-evaluations"
+import {useLoadTestsetsList} from "@/services/testsets"
 import {isDemo} from "@/lib/helpers/utils"
 import {Button, Col, Dropdown, MenuProps, Modal, ModalProps, Row, Spin, message} from "antd"
 import {getErrorMessage} from "@/lib/helpers/errorHandler"
@@ -11,7 +13,6 @@ import {getAllVariantParameters} from "@/lib/helpers/variantHelper"
 import {useRouter} from "next/router"
 import {useAppTheme} from "../Layout/ThemeContextProvider"
 import {createUseStyles} from "react-jss"
-import {getApikeys} from "@/lib/helpers/llmProviders"
 import EvaluationErrorModal from "../Evaluations/EvaluationErrorModal"
 import {dynamicComponent} from "@/lib/helpers/dynamic"
