+++ agenta-web/src/components/Playground/LoadTestsModal.tsx
@@ -1,5 +1,5 @@
-import {loadTestset, useLoadTestsetsList} from "@/lib/services/api"
-import {Button, Divider, Dropdown, Modal, Select, Space} from "antd"
+import {fetchTestset, useLoadTestsetsList} from "@/services/testsets"
+import {Button, Divider, Modal, Select} from "antd"
 import {useRouter} from "next/router"
 import {PropsWithChildren, useState} from "react"
 import {createUseStyles} from "react-jss"
@@ -38,7 +38,7 @@ const LoadTestsModal: React.FC<Props> = (props) => {
     }))
 
     const handleClick = (shouldReplace: boolean) => {
-        loadTestset(selectedSet).then((data) => {
+        fetchTestset(selectedSet).then((data) => {
             onLoad(data.csvdata, shouldReplace)
         })
         setIsOpen(false)
