+++ agenta-web/src/components/Playground/ViewNavigation.tsx
@@ -9,12 +9,12 @@ import {useState} from "react"
 import axios from "axios"
 import {createUseStyles} from "react-jss"
 import {
+    fetchAppContainerURL,
+    deleteSingleVariant,
     fetchVariantLogs,
-    getAppContainerURL,
-    removeVariant,
     restartAppVariantContainer,
     waitForAppToStart,
-} from "@/lib/services/api"
+} from "@/services/api"
 import {useAppsData} from "@/contexts/app.context"
 import {isDemo} from "@/lib/helpers/utils"
 import ResultComponent from "../ResultComponent/ResultComponent"
@@ -141,7 +141,7 @@ const ViewNavigation: React.FC<Props> = ({
         }
 
         const variantContainerPath = async () => {
-            const url = await getAppContainerURL(appId, variant.variantId, variant.baseId)
+            const url = await fetchAppContainerURL(appId, variant.variantId, variant.baseId)
             setContainerURI(url)
         }
         if (!containerURI) {
@@ -231,7 +231,7 @@ const ViewNavigation: React.FC<Props> = ({
                             type="primary"
                             danger
                             onClick={() => {
-                                deleteVariant(() => removeVariant(variant.variantId))
+                                deleteVariant(() => deleteSingleVariant(variant.variantId))
                             }}
                         >
                             <Tooltip placement="bottom" title="Delete the variant permanently">
