+++ agenta-web/src/components/Playground/Views/ParametersView.tsx
@@ -5,16 +5,17 @@ import React, {useEffect, useState} from "react"
 import {createUseStyles} from "react-jss"
 import {ModelParameters, ObjectParameters, StringParameters} from "./ParametersCards"
 import PublishVariantModal from "./PublishVariantModal"
-import {promptVersioning, removeVariant} from "@/lib/services/api"
+import {deleteSingleVariant} from "@/services/api"
 import {CloudUploadOutlined, DeleteOutlined, HistoryOutlined, SaveOutlined} from "@ant-design/icons"
 import {usePostHogAg} from "@/hooks/usePostHogAg"
 import {isDemo} from "@/lib/helpers/utils"
 import {useQueryParam} from "@/hooks/useQuery"
-import {dynamicComponent} from "@/lib/helpers/dynamic"
+import {dynamicComponent, dynamicService} from "@/lib/helpers/dynamic"
 
 const PromptVersioningDrawer: any = dynamicComponent(
     `PromptVersioningDrawer/PromptVersioningDrawer`,
 )
+const promptVersioning: any = dynamicService("promptVersioning")
 
 interface Props {
     variant: Variant
@@ -133,7 +134,7 @@ const ParametersView: React.FC<Props> = ({
     const handleDelete = () => {
         deleteVariant(() => {
             if (variant.persistent) {
-                return removeVariant(variant.variantId).then(() => {
+                return deleteSingleVariant(variant.variantId).then(() => {
                     onStateChange(false)
                 })
             }
@@ -151,10 +152,13 @@ const ParametersView: React.FC<Props> = ({
         setHistoryStatus({loading: true, error: false})
         setIsDrawerOpen(true)
         try {
-            if (variant.variantId && isDemo()) {
-                const revisions = await promptVersioning(variant.variantId)
-                setPromptRevisions(revisions)
-            }
+            await promptVersioning.then(async (module: any) => {
+                if (!module) return
+                if (variant.variantId && isDemo()) {
+                    const revisions = await module.fetchAllPromptVersioning(variant.variantId)
+                    setPromptRevisions(revisions)
+                }
+            })
             setHistoryStatus({loading: false, error: false})
         } catch (error) {
             setHistoryStatus({loading: false, error: true})
