+++ agenta-web/src/components/TestSetTable/TestsetTable.tsx
@@ -4,7 +4,8 @@ import {createUseStyles} from "react-jss"
 import {Button, Input, Tooltip, Typography, message} from "antd"
 import TestsetMusHaveNameModal from "./InsertTestsetNameModal"
 import {DeleteOutlined, EditOutlined, PlusOutlined} from "@ant-design/icons"
-import {createNewTestset, fetchVariants, loadTestset, updateTestset} from "@/lib/services/api"
+import {fetchVariants} from "@/services/api"
+import {createNewTestset, fetchTestset, updateTestset} from "@/services/testsets"
 import {useRouter} from "next/router"
 import {useAppTheme} from "../Layout/ThemeContextProvider"
 import useBlockNavigation from "@/hooks/useBlockNavigation"
@@ -207,7 +208,7 @@ const TestsetTable: React.FC<testsetTableProps> = ({mode}) => {
 
         if (mode === "edit" && testset_id) {
             setLoading(true)
-            loadTestset(testset_id as string).then((data) => {
+            fetchTestset(testset_id as string).then((data) => {
                 setTestsetName(data.name)
                 setRowData(data.csvdata)
                 applyColData(
