+++ agenta-web/src/contexts/profile.context.tsx
@@ -2,7 +2,7 @@ import {usePostHogAg} from "@/hooks/usePostHogAg"
 import {useSession} from "@/hooks/useSession"
 import useStateCallback from "@/hooks/useStateCallback"
 import {isDemo} from "@/lib/helpers/utils"
-import {getProfile} from "@/lib/services/api"
+import {fetchProfile} from "@/services/api"
 import {User} from "@/lib/Types"
 import {PropsWithChildren, createContext, useState, useContext, useEffect, useCallback} from "react"
 
@@ -36,7 +36,7 @@ const ProfileContextProvider: React.FC<PropsWithChildren> = ({children}) => {
 
     const fetcher = useCallback((onSuccess?: () => void) => {
         setLoading(true)
-        getProfile()
+        fetchProfile()
             .then((profile) => {
                 posthog.identify()
                 setUser(profile.data, onSuccess)
