+++ agenta-web/src/services/evaluations/index.ts
@@ -21,9 +21,7 @@ import webhookImg from "@/media/link.png"
 import aiImg from "@/media/artificial-intelligence.png"
 import codeImg from "@/media/browser.png"
 import bracketCurlyImg from "@/media/bracket-curly.png"
-import dayjs from "dayjs"
-import {loadTestset} from "@/lib/services/api"
-import {runningStatuses} from "@/components/pages/evaluations/cellRenderers/cellRenderers"
+import {fetchTestset} from "@/services/testsets"
 import {calcEvalDuration} from "@/lib/helpers/evaluate"
 import _ from "lodash"
 
@@ -213,8 +211,7 @@ export const updateAnnotationScenario = async (
 // Comparison
 export const fetchAllComparisonResults = async (evaluationIds: string[]) => {
     const scenarioGroups = await Promise.all(evaluationIds.map(fetchAllEvaluationScenarios))
-    const testset: TestSet = await loadTestset(scenarioGroups[0][0].evaluation?.testset?.id)
-
+    const testset: TestSet = await fetchTestset(scenarioGroups[0][0].evaluation?.testset?.id)
     const inputsNameSet = new Set<string>()
     scenarioGroups.forEach((group) => {
         group.forEach((scenario) => {
